pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jira-mcp-server"
version = "1.0.0"
description = "A clean, reliable MCP server for Jira integration with Claude Desktop"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Ben Vierck", email = "ben@positronic.ai"}
]
maintainers = [
    {name = "Ben Vierck", email = "ben@positronic.ai"}
]
keywords = [
    "jira",
    "mcp",
    "model-context-protocol",
    "claude",
    "anthropic",
    "api",
    "atlassian"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Positronic-AI/jira-mcp"
Documentation = "https://github.com/Positronic-AI/jira-mcp#readme"
Repository = "https://github.com/Positronic-AI/jira-mcp"
Issues = "https://github.com/Positronic-AI/jira-mcp/issues"
Changelog = "https://github.com/Positronic-AI/jira-mcp/blob/main/CHANGELOG.md"

[project.scripts]
jira-mcp = "jira_mcp.server:cli"

[tool.setuptools]
packages = ["jira_mcp"]

[tool.setuptools.package-data]
jira_mcp = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
